+++ R/pk.calc.all.R
@@ -322,12 +322,16 @@ pk.nca.intervals <- function(data_conc, data_dose, data_intervals, sparse,
 #'   for urine and fecal measurements)
 #' @param dose,dose.group Dose amount (may be a scalar or vector) for the
 #'   current interval or all data for the group
-#' @param time.dose,time.dose.group Time of the dose for the current interval or
-#'   all data for the group (must be the same length as `dose` or `dose.group`)
-#' @param duration.dose,duration.dose.group The duration of the dose
-#'   administration for the current interval or all data for the group
-#'   (typically zero for extravascular and intravascular bolus and nonzero for
-#'   intravascular infusion)
+#' @param time.dose Time of the dose for the current interval (must be the same
+#'   length as `dose`)
+#' @param time.dose.group Time of the dose for all data for the group (must be
+#'   the same length as `dose.group`)
+#' @param duration.dose The duration of the dose administration for the current
+#'   interval (typically zero for extravascular and intravascular bolus and
+#'   nonzero for intravascular infusion)
+#' @param duration.dose.group The duration of the dose administration for all
+#'   data for the group (typically zero for extravascular and intravascular
+#'   bolus and nonzero for intravascular infusion)
 #' @param route,route.group The route of dosing for the current interval or all
 #'   data for the group
 #' @param impute_method The method to use for imputation as a character string