+++ NEWS.md
@@ -10,6 +10,7 @@ the dosing including dose amount and route.
 
 * Units for fraction excretion parameter (fe) are now accurately captured as
   amount/dose units rather than "fraction" (#426)
+* `get_halflife_points` will disconsider later points to `lambda.z.time.last`, instead of `tlast` (#448)
 * `lambda.z` calculations will now only consider time points that occur after
   the end of the latest dose administration (#139)
 * `aucint.inf.pred` is `NA` when half-life is not estimable (#450)